Move ShowIssuesList to shared
